+++ canvas/packages/excalidraw/components/CommandPalette/CommandPalette.tsx
@@ -369,7 +369,6 @@ function CommandPaletteInner({
         actionManager.actions.selectAll,
         actionManager.actions.toggleElementLock,
         actionManager.actions.unlockAllElements,
-        actionManager.actions.stats,
       ].map((action) => actionToCommand(action, DEFAULT_CATEGORIES.editor));
 
       const exportCommands: CommandPaletteItem[] = [